More type checking - Split typings in seperate files. - Validate environment variables on start. - Added error handling. - Remove caching of guilds as this was tedious to keep up to date. - Reduce nesting in `success` command and cronjob creation. - Lower max arguments for the `examen` command to 10. - Update the regex in the `examen` command to find years too. - Update `info` command:   - Generate bot invite.   - Get bot version from process variable.   - Display discord.js and node versions.   - Display link to support server. - Remove all guild references when a bot leaves a guild. - Update interactionCreate:   - Check if interaction was send in a guild by an user.   - Check if interaction exists.   - Use built-in `splitMessage`. - Update command handler:   - Better handling if there is an error with a command.   - More checks for the commands.   - Move creation of slash commands to the command handler. - Store defaults to database. - Remove permission and config file.

+++ src/events/guild/guildDelete.js
@@ -1,6 +1,7 @@
 const { Guild } = require('discord.js');
-const { BotClient } = require('../../typings');
-const guildModel = require('../../models/guildModel');
+const { guildModel } = require('../../models/guildModel');
+const { profileModel } = require('../../models/profileModel');
+const BotClient = require('../../structures/BotClient');
 
 /**
  * @param {BotClient} client
@@ -8,12 +9,26 @@ const guildModel = require('../../models/guildModel');
  */
 module.exports = async (client, guild) => {
   try {
+    if (!guild) return;
+
     // Delete a database profile when the bot leaves a guild or if the guild gets deleted
     await guildModel.findOneAndDelete({ _id: guild.id });
 
-    // Remove the guild from our cache
-    client.guildInfo.delete(guild.id);
+    // Delete all the references to the guild in user.cooldowns
+    await profileModel.updateMany(
+      { },
+      { $pull: { cooldowns: { guildId: guild.id } } },
+    );
+
+    // Delete all the references to the guild in user.exams
+    await profileModel.updateMany(
+      { },
+      { $pull: { exams: { guildId: guild.id } } },
+    );
+
+    // Delete the cronjob that runs in this guild
+    client.cronJobs.delete(guild.id);
   } catch (error) {
-    client.log.error('Er is een fout opgetreden bij het verwijderen van een database profiel voor een server.', error);
+    client.log.error('Er is een fout opgetreden bij het verwijderen van een gilde uit de database.', error);
   }
 };
